Refactor autoProcessComics -> comics
