+++ core/__init__.py
@@ -43,7 +43,7 @@ import six
 from six.moves import reload_module
 
 from core import logger, nzbToMediaDB, versionCheck
-from core.auto_process.autoProcessComics import Comic
+from core.auto_process.comics import Comic
 from core.auto_process.autoProcessGames import Game
 from core.auto_process.autoProcessMovie import Movie
 from core.auto_process.autoProcessMusic import Music
