+++ core/auto_process/comics.py
