+++ nzbToMedia.py
@@ -629,7 +629,7 @@ import sys
 
 import core
 from core import logger, nzbToMediaDB
-from core.auto_process.autoProcessComics import Comic
+from core.auto_process.comics import Comic
 from core.auto_process.autoProcessGames import Game
 from core.auto_process.autoProcessMovie import Movie
 from core.auto_process.autoProcessMusic import Music
